Cut over swift-sheet-music's in-tree wire-format toolkit to the published
swift-wirelet package at v0.1.0-alpha.2 (revision 31be47c).

Swift side:
- Add `.package(url: "git@github.com:jiyimeta/swift-wirelet.git", revision: …)` SwiftPM dep (private repo; SSH form per spec).
- Delete 5 in-tree Source dirs (SheetMusicWireFormat, SheetMusicWireFormatMacros, WireFormatSchema, WireFormatKotlinEmitter, EmitKotlinCodecs) and 3 corresponding test dirs.
- Mass-rename `import SheetMusicWireFormat` → `import Wirelet` across 27 source + 9 test files.
- Drop swift-syntax dep (only the deleted in-tree macros needed it).
- Reshuffle Sources/SheetMusicAndroidJNI/ into per-codec-package sub-dirs (Metadata/, Audio/, Draw/) so each Gradle module's wirelet { sources } scans a coherent dir.

Android side:
- Apply `io.github.jiyimeta.wirelet` Gradle plugin to SheetMusicAndroid, SheetMusicAudioAndroid, and Examples/Android/app — replaces the legacy `Exec` tasks calling `swift run emit-kotlin-codecs`.
- Add `api("io.github.jiyimeta:wirelet-runtime:0.1.0-alpha.2")` dep; delete hand-written BinaryReader/Writer.kt under sheetmusic.wireformat.
- Rename 10 Kotlin files from `io.github.jiyimeta.sheetmusic.wireformat` → `io.github.jiyimeta.wirelet`.
- Add WireletGitHubPackages maven block to both Android/ and Examples/Android/ settings.gradle.kts (env-var-first creds with gpr.* fallback; includeGroupByRegex covers both runtime + plugin marker POM).
- Fix two Kotlin production sites where array decoding still used the legacy positional layout (AndroidPlaybackEngine.kt + GMInstrumentDecoder.kt).
- /wirelet-checkout symlink gitignored; CLAUDE.md documents the dev-override bootstrap.

Tests:
- Regenerate 13 golden binary fixtures under Tests/SheetMusicTests/Resources/Golden/Audio/ in wirelet TLV format.
- Delete WireFormatTests.swift entirely (duplicates wirelet's own primitive + macro tests).
- Drop hardcoded byte-sequence assertions from 11 codec test suites (round-trip + golden binaries now cover the cross-language byte contract).

Verifications green:
- swift build + swift test (1331 tests passed)
- xcodebuild SheetMusicExample (iOS Simulator) + SheetMusicExampleMac
- cd Android && ./gradlew clean assembleRelease + test
- cd Examples/Android && ./gradlew :app:assembleDebug

+++ Examples/Android/app/src/main/java/com/example/sheetmusic/draw/model/DrawCommand.kt
@@ -16,7 +16,7 @@ sealed class DrawCommand {
         val w: Double, val h: Double,
     ) : DrawCommand()
     data class Glyph(
-        val codepoint: Long,
+        val codepoint: UInt,
         val x: Double, val y: Double,
         val size: Double,
         val fontId: FontID,
@@ -28,7 +28,7 @@ sealed class DrawCommand {
         val fontId: FontID,
     ) : DrawCommand()
     /** Active paint colour as packed ARGB (0xAARRGGBB). */
-    data class SetColor(val argb: Long) : DrawCommand()
+    data class SetColor(val argb: UInt) : DrawCommand()
     /** Cubic Bezier curve from current path point to (x, y). */
     data class CubicTo(
         val cx1: Double, val cy1: Double,
